update ws
Former-commit-id: 5a3755583cc032818e93fdb9d2c0b19d41940246

+++ org.getalp.lexsema-ws/src/main/java/org/getalp/lexsema/ws/w2v/Word2VecWebService.java
@@ -199,7 +199,7 @@ public class Word2VecWebService extends WebServiceServlet
             wordsIndexes = new HashMap<>();
             vectors = new double[nbWords][vectorDimension];
             for (int i = 0 ; i < nbWords ; i++) {
-                int percentage =  (((i + 1) / nbWords) * 100);
+                int percentage = ((int) ((((double) (i + 1)) / ((double) (nbWords))) * 100.0));
                 System.out.print("Adding words... (" + percentage + "%)\r");
                 words[i] = w2v.vocab().wordAtIndex(i);
                 wordsIndexes.put(words[i], i);
